+++ src/bract/core/type.clj
@@ -10,12 +10,12 @@
 (ns bract.core.type)
 
 
-(defprotocol Inducer
+(defprotocol IFunction
   (ifunc [this] [this config-key] "Return the inducer function to be invoked")
   (iname [this] "Return the name (or derived name) of the inducer")
   (iargs [this] "Return the additional arguments to be passed to the inducer function"))
 
 
-(defn inducer?
+(defn function?
   [x]
-  (satisfies? Inducer x))
+  (satisfies? IFunction x))